+++ src/plc_read_db.h
@@ -1,0 +1,11 @@
+#ifndef _PLC_READ_DB_H_
+#define _PLC_READ_DB_H_
+
+typedef struct
+{
+    int db_index;
+
+}PLC_ReadDB;
+
+
+#endif